Refactoring.
Moving package net.finmath.experimental.model.implementation.* to
net.finmath.modelling.descriptor.*
