+++ src/main/java/net/finmath/modelling/descriptor/SingleAssetEuropeanOptionProductDescriptor.java
@@ -4,7 +4,7 @@
  * Created on 09.02.2018
  */
 
-package net.finmath.experimental.model.implementation;
+package net.finmath.modelling.descriptor;
 
 import net.finmath.modelling.SingleAssetProductDescriptor;
