+++ src/main/java/net/finmath/modelling/descriptor/SingleAssetMonteCarloProductFactory.java
@@ -4,7 +4,7 @@
  * Created on 09.02.2018
  */
 
-package net.finmath.experimental.model.implementation;
+package net.finmath.modelling.descriptor;
 
 import net.finmath.modelling.Product;
 import net.finmath.modelling.ProductFactory;
@@ -13,19 +13,19 @@ import net.finmath.modelling.SingleAssetProductDescriptor;
 /**
  * @author Christian Fries
  */
-public class SingleAssetFourierProductFactory implements ProductFactory<SingleAssetProductDescriptor> {
+public class SingleAssetMonteCarloProductFactory implements ProductFactory<SingleAssetProductDescriptor> {
 
 	/**
 	 * Create factory.
 	 */
-	public SingleAssetFourierProductFactory() {
+	public SingleAssetMonteCarloProductFactory() {
 	}
 
 	@Override
 	public Product<?> getProductFromDescription(SingleAssetProductDescriptor descriptor) {
 
 		if(descriptor instanceof SingleAssetEuropeanOptionProductDescriptor) {
-			Product<SingleAssetEuropeanOptionProductDescriptor> product = new net.finmath.fouriermethod.products.EuropeanOption((SingleAssetEuropeanOptionProductDescriptor) descriptor);
+			Product<SingleAssetEuropeanOptionProductDescriptor> product = new net.finmath.montecarlo.assetderivativevaluation.products.EuropeanOption((SingleAssetEuropeanOptionProductDescriptor) descriptor);
 			return product;
 		}
 		else {
